Fully revamped the command system: - Removed any trace of the old ginterface system - Reverted back to a simpler argument based system - Added support for permissions Implemented Paginations Temporarily disabled SpamHandler Refactored HastebinScanner Changed single variables in Logger to lists. Faqs: - Added a Faq data object containing the faq key, text and author ID. - Made add faq, edit faq, get faq, list faqs, and remove faq commands. - Added table into mysql schema for faq storage Added QuestionnaireUtils class with method to ask a question and get a response. Renamed me.piggypiglet.gary.storage.json to me.piggypiglet.gary.storage.file Added RoleUtils class to test if a member is part of a role or any roles higher. Made MySQLUtils methods return a boolean based on query success Added Language system Added embedded file for version usage Added a slowmode command

+++ src/main/java/me/piggypiglet/gary/core/handlers/misc/MySQLHandler.java
@@ -4,6 +4,7 @@ import me.piggypiglet.gary.core.handlers.GEvent;
 import me.piggypiglet.gary.core.objects.enums.EventsEnum;
 import me.piggypiglet.gary.core.utils.mysql.MessageUtils;
 import net.dv8tion.jda.core.events.Event;
+import net.dv8tion.jda.core.events.message.guild.GenericGuildMessageEvent;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageDeleteEvent;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageReceivedEvent;
 import net.dv8tion.jda.core.events.message.guild.GuildMessageUpdateEvent;
@@ -21,20 +22,24 @@ public final class MySQLHandler extends GEvent {
 
     @Override
     protected void execute(Event event) {
-        switch (EventsEnum.fromEvent(event)) {
-            case MESSAGE_CREATE:
-                MessageUtils.addMessage(((GuildMessageReceivedEvent) event).getMessage());
-                break;
+        GenericGuildMessageEvent e = (GenericGuildMessageEvent) event;
 
-            case MESSAGE_EDIT:
-                MessageUtils.editMessage(((GuildMessageUpdateEvent) event).getMessage());
-                break;
+        if (!MessageUtils.getAuthor(e.getMessageIdLong()).isBot()) {
+            switch (EventsEnum.fromEvent(event)) {
+                case MESSAGE_CREATE:
+                    MessageUtils.addMessage(((GuildMessageReceivedEvent) event).getMessage());
+                    break;
 
-            case MESSAGE_DELETE:
-                // sleep before deleting from mysql so logger can grab the message content
-                instance.sleep(100);
-                MessageUtils.deleteMessage(((GuildMessageDeleteEvent) event).getMessageIdLong());
-                break;
+                case MESSAGE_EDIT:
+                    MessageUtils.editMessage(((GuildMessageUpdateEvent) event).getMessage());
+                    break;
+
+                case MESSAGE_DELETE:
+                    // sleep before deleting from mysql so logger can grab the message content
+                    instance.sleep(100);
+                    MessageUtils.deleteMessage(((GuildMessageDeleteEvent) event).getMessageIdLong());
+                    break;
+            }
         }
     }
 }
